API

Logging Levels

LoggingCommon.AboveMax — Constant

AboveMax

Alias for LogLevel(2147483646)::Base.Logging.LogLevel

LoggingCommon.Alert — Constant

Alert

Alias for LogLevel(2010)::Base.Logging.LogLevel

LoggingCommon.All — Constant

All

Alias for LogLevel(-2147483647)::Base.Logging.LogLevel

LoggingCommon.Critical — Constant

Critical

Alias for LogLevel(2001)::Base.Logging.LogLevel

LoggingCommon.Emergency — Constant

Emergency

Alias for LogLevel(2100)::Base.Logging.LogLevel

LoggingCommon.Fatal — Constant

Fatal

Alias for LogLevel(2147483645)::Base.Logging.LogLevel

LoggingCommon.NotSet — Constant

NotSet

Alias for LogLevel(-2147483648)::Base.Logging.LogLevel

LoggingCommon.Notice — Constant

Notice

Alias for LogLevel(1)::Base.Logging.LogLevel

LoggingCommon.Off — Constant

Off

Alias for LogLevel(2147483647)::Base.Logging.LogLevel

LoggingCommon.Trace — Constant

Trace

Alias for LogLevel(-1001)::Base.Logging.LogLevel

LoggingCommon.NamedLogLevel — Type

NamedLogLevel(name::Symbol)

Represents a logging level with an associated name

LoggingCommon.is_valid_log_level — Method

is_valid_log_level(level::Symbol) -> Bool

Returns true if level is a known log level

LoggingCommon.log_level — Function

log_level(level) -> LogLevel

Converts the input level to a LogLevel type

LoggingCommon.log_level_name — Method

log_level_name(level) -> String

Returns the name associated with level as a String

LoggingCommon.nearest_log_level — Method

nearest_log_level(input::LogLevel) -> Symbol

Returns the first named log level L with log_level(L) ≥ input

LoggingCommon.validate_log_level — Method

validate_log_level(level::Symbol)

Throws an UnknownLogLevelException if level is not a known named log level. Otherwise, returns nothing.

Logging Records

LoggingCommon.AbstractLogRecord — Type

AbstractLogRecord

Abstract supertype of all log records

LoggingCommon.LogRecord — Type

LogRecord(static_meta, runtime_meta, record, data)

LogRecord(static_meta, runtime_meta, record, args...)

LogRecord(static_meta, record, args...; runtime_meta=RuntimeLogRecordMetadata())

A log record with an associated StaticLogRecordMetadata, RuntimeLogRecordMetadata, LogRecordData and an underlying record. If data is not provided, it is constructed from args....

Arguments

static_meta::StaticLogRecordMetadata - Static log record metadata

runtime_meta::RuntimeLogRecordMetadata - Runtime log record metadata

record::AbstractLogRecord - Underlying record

data::LogRecordData - Optional key => value pairs attached to record. Constructed from args... if not provided.

LoggingCommon.LogRecordData — Type

LogRecordData(data)

LogRecordData(args::Pair{<:Any, <:Any}...)

A type representing an optional collection of key => value pairs attached to a log record.

data must be an iterable collection where each element is a Pair

LoggingCommon.MessageLogRecord — Type

MessageLogRecord(message::AbstractString)

A type representing a log record with an associated message

LoggingCommon.RuntimeLogRecordMetadata — Type

RuntimeLogRecordMetadata(datetime::DateTime, thread_id::Int, worker_id::Int)

RuntimeLogRecordMetadata(; datetime=Dates.now(), thread_id::Int=Threads.threadid(), worker_id::Int=_worker_id())

Components of a generic log record which are known at runtime

LoggingCommon.StacktraceLogRecord — Type

StacktraceLogRecord(staticmeta, runtimemeta, exception, stacktrace, data) StacktraceLogRecord(staticmeta, stacktrace, data; runtimemeta=RuntimeLogRecordMetadata(), exception=nothing) StacktraceLogRecord(staticmeta, stacktrace, args...; runtimemeta=RuntimeLogRecordMetadata(), exception=nothing)

An AbstractLogRecord type with an associated stacktrace and optional exception and a collection of key-value data.

Arguments

static_meta::StaticLogRecordMetadata - Static log record metadata

runtime_meta[_metadata]::RuntimeLogRecordMetadata - Runtime log record metadata

exception::Union{Nothing,Exception} - Exception for log record

stacktrace::Base.StackTraces.StackTrace - Stacktrace for log record

data::LogRecordData - Optional key => value pairs attached to record. Constructed from args... if not provided.

LoggingCommon.StaticLogRecordMetadata — Type

StaticLogRecordMetadata(source::String, level::LogLevel, level_name::String, filename::Union{String,Nothing}, line_num::Int [, group=nothing] [, id=nothing])

Components of a generic log record that are known at compile time

LoggingCommon.add_record_data! — Method

add_record_data!(r, data::Pair)

Adds the data := key => value pair to r

LoggingCommon.is_error_record — Method

is_error_record(record)

Returns true if record is a log record associated with an error/exception, false otherwise.

LoggingCommon.log_record_data — Method

log_record_data(record)

Returns an iterator over the key => value pairs associated with record

LoggingCommon.log_record_data — Method

log_record_data(kv_pairs; [exclude=()]) -> LogRecordData

Returns a LogRecordData from the input key => value pairs

LoggingCommon.log_record_data — Method

log_record_data() -> LogRecordData{Symbol}

LoggingCommon.message_log_record — Method

message_log_record(static_meta::StaticLogRecordMetadata, message::AbstractString, data; [runtime_meta = RuntimeLogRecordMetadata()])

LoggingCommon.runtime_metadata — Method

runtime_metadata(record)

Returns the RuntimeLogRecordMetadata associated with record or nothing if it is not present

LoggingCommon.static_metadata — Method

static_metadata(record)

Returns the StaticLogRecordMetadata associated with record or nothing if it is not present